hw3/booking-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.31.1
email-validator==2.2.0